requirements.txt
flask==2.3.3
flask-cors>=4.0.0
openai-whisper==20230314
numpy>=1.24.0
moviepy==1.0.3
Pillow>=9.5.0
werkzeug>=2.3.7
rich>=13.0.0
termcolor>=2.3.0
librosa>=0.10.0
soundfile>=0.12.1
ultralytics>=8.0.0
pyannote.audio>=3.0.0
torch>=2.0.0
torchaudio>=2.0.0
opencv-python>=4.8.0
pydub>=0.25.1
asteroid>=0.6.0
asteroid-filterbanks>=0.4.0
torchmetrics<=0.11.4
packaging<25.0,>=20.0
lightning<2.6.0
scipy>=1.7.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
boto3>=1.28.0
botocore>=1.31.0